--- drivers/staging/zsmalloc/Makefile
@@ -1,3 +1,0 @@
-zsmalloc-y 		:= zsmalloc-main.o
-
-obj-$(CONFIG_ZSMALLOC)	+= zsmalloc.o
