+++ mm/Makefile
@@ -53,3 +53,4 @@ obj-$(CONFIG_DEBUG_KMEMLEAK) += kmemleak.o
 obj-$(CONFIG_DEBUG_KMEMLEAK_TEST) += kmemleak-test.o
 obj-$(CONFIG_CLEANCACHE) += cleancache.o
 obj-$(CONFIG_PANTECH_MEM_LEAK_TRACE) += pantech_memleak.o
+obj-$(CONFIG_ZSMALLOC)	+= zsmalloc.o
